+++ js/pages/suggest.js
@@ -2,7 +2,7 @@
 
 import * as store from '../store.js';
 import * as engine from '../engine.js';
-import { resolveImageUri } from '../tauri-fs.js';
+import { resolveImageUri } from '../electron-bridge.js';
 
 function esc(t) {
   return String(t ?? '').replace(/[&<>"']/g, m => ({'&':'&amp;','<':'&lt;','>':'&gt;','"':'&quot;',"'":'&#039;'}[m]));